Add DirtyCheck annotation and change owner functionality
The DirtyCheck annotation has been added to the AbstractAclObjectIdentity class to improve change tracking. Additionally, a 'change owner' function with corresponding integration tests has been added to the AclUtilServiceSpec class to allow for changing the owner of an object. The tests ensure that the change in ownership is properly reflected in both Acl and persistent data.